pyproject.toml
[tool.poetry]
name = "web-mod-10"
version = "0.1.0"
description = ""
authors = ["OleksandrSkuz1 <138138803+OleksandrSkuz1@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
setuptools = "^69.5.1"
django = "^5.0.5"
pymongo = "^4.7.2"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
